requirements.txt
streamlit
torch
torchvision
Pillow
matplotlib
seaborn